--- CheahMunXi/chat/.env.example
@@ -1,22 +1,0 @@
-APP_NAME=Lost And Found MMU Chat App
-SECRET_KEY=my_super_secret_random_key_123456
-HOST=localhost
-PORT=8080
-
-DB_CONNECTION=sqlite
-DB_DATABASE=first.db
-
-MAIL_MAILER=smtp
-MAIL_HOST=smtp.gmail.com
-MAIL_PORT=587
-MAIL_USERNAME=your_email@gmail.com
-MAIL_PASSWORD=replace_with_google_app_password
-MAIL_FROM_ADDRESS=your_email@gmail.com
-MAIL_ENCRYPTION=tls
-
-EMAIL_NOTIFICATIONS_ENABLED=FALSE
-EMAIL_NOTIFICATIONS_COOLDOWN=600
-
-DEBUG=TRUE
-LOG_LEVEL=debug
-LOG_FILE=app.log
