--- CheahMunXi/chat/.nicegui/storage-user-fa022289-e8df-4046-be6a-9f990745c59d.json
@@ -1,1 +1,0 @@
-{"authenticated":true,"username":"munxi","user_id":1}
