--- CheahMunXi/chat/handlers/__init__.py
@@ -1,14 +1,0 @@
-"""
-Handlers package
-"""
-
-from .auth import handle_login, handle_register, handle_logout
-from .chat import send_message, open_chat
-
-__all__ = [
-    'handle_login',
-    'handle_register', 
-    'handle_logout',
-    'send_message',
-    'open_chat'
-]
