+++ CheahMunXi/chat_v2/.env.example
@@ -1,0 +1,22 @@
+APP_NAME=First Web App
+SECRET_KEY=
+HOST=localhost
+PORT=8080
+
+DB_CONNECTION=sqlite
+DB_DATABASE=first.db
+
+MAIL_MAILER=smtp
+MAIL_HOST=smtp.gmail.com
+MAIL_PORT=587
+MAIL_USERNAME=your_email@gmail.com
+MAIL_PASSWORD=your_app_password
+MAIL_FROM_ADDRESS=your_email@gmail.com
+MAIL_ENCRYPTION=tls
+
+EMAIL_NOTIFICATIONS_ENABLED=TRUE
+EMAIL_NOTIFICATIONS_COOLDOWN=600
+
+DEBUG=FALSE
+LOG_LEVEL=info
+LOG_FILE=app.log
