+++ CheahMunXi/chat_v2/handlers/__init__.py
@@ -1,0 +1,13 @@
+"""
+Handlers package
+"""
+
+from .auth import handle_login, handle_register, handle_logout
+from .chat import send_message
+
+__all__ = [
+    'handle_login',
+    'handle_register', 
+    'handle_logout',
+    'send_message',
+]
